pkg/cover/backend: add mach-o object support for darwin
I have invested time deduplicating code between the two formats. As the
go modules for Mach-O and ELF don't share types, I had to re-wrap some
information or move code to format specific functions though.
